+++ src/getStepStartStates.test.ts
@@ -34,11 +34,7 @@ import {
   STEP_RM_NETWORK_SECURITY_GROUP_RULE_RELATIONSHIPS,
   STEP_RM_NETWORK_AZURE_FIREWALLS,
 } from './steps/resource-manager/network';
-import {
-  STEP_RM_STORAGE_RESOURCES,
-  STEP_RM_STORAGE_QUEUES,
-  STEP_RM_STORAGE_TABLES,
-} from './steps/resource-manager/storage';
+import { steps as storageSteps } from './steps/resource-manager/storage';
 import { IntegrationConfig } from './types';
 import {
   STEP_RM_AUTHORIZATION_ROLE_ASSIGNMENTS,
@@ -132,9 +128,9 @@ describe('getStepStartStates', () => {
       [STEP_RM_DATABASE_MYSQL_DATABASES]: { disabled: true },
       [STEP_RM_DATABASE_SQL_DATABASES]: { disabled: true },
       [STEP_RM_DATABASE_POSTGRESQL_DATABASES]: { disabled: true },
-      [STEP_RM_STORAGE_RESOURCES]: { disabled: true },
-      [STEP_RM_STORAGE_QUEUES]: { disabled: true },
-      [STEP_RM_STORAGE_TABLES]: { disabled: true },
+      [storageSteps.STORAGE_RESOURCES]: { disabled: true },
+      [storageSteps.STORAGE_QUEUES]: { disabled: true },
+      [storageSteps.STORAGE_TABLES]: { disabled: true },
       [STEP_RM_COMPUTE_NETWORK_RELATIONSHIPS]: { disabled: true },
       [STEP_RM_AUTHORIZATION_ROLE_ASSIGNMENTS]: { disabled: true },
       [STEP_RM_AUTHORIZATION_ROLE_ASSIGNMENT_PRINCIPAL_RELATIONSHIPS]: {
@@ -209,9 +205,9 @@ describe('getStepStartStates', () => {
       [STEP_RM_DATABASE_MYSQL_DATABASES]: { disabled: true },
       [STEP_RM_DATABASE_SQL_DATABASES]: { disabled: true },
       [STEP_RM_DATABASE_POSTGRESQL_DATABASES]: { disabled: true },
-      [STEP_RM_STORAGE_RESOURCES]: { disabled: true },
-      [STEP_RM_STORAGE_QUEUES]: { disabled: true },
-      [STEP_RM_STORAGE_TABLES]: { disabled: true },
+      [storageSteps.STORAGE_RESOURCES]: { disabled: true },
+      [storageSteps.STORAGE_QUEUES]: { disabled: true },
+      [storageSteps.STORAGE_TABLES]: { disabled: true },
       [STEP_RM_COMPUTE_NETWORK_RELATIONSHIPS]: { disabled: true },
       [STEP_RM_AUTHORIZATION_ROLE_ASSIGNMENTS]: { disabled: true },
       [STEP_RM_AUTHORIZATION_ROLE_ASSIGNMENT_PRINCIPAL_RELATIONSHIPS]: {
@@ -286,9 +282,9 @@ describe('getStepStartStates', () => {
       [STEP_RM_DATABASE_MYSQL_DATABASES]: { disabled: false },
       [STEP_RM_DATABASE_SQL_DATABASES]: { disabled: false },
       [STEP_RM_DATABASE_POSTGRESQL_DATABASES]: { disabled: false },
-      [STEP_RM_STORAGE_RESOURCES]: { disabled: false },
-      [STEP_RM_STORAGE_QUEUES]: { disabled: false },
-      [STEP_RM_STORAGE_TABLES]: { disabled: false },
+      [storageSteps.STORAGE_RESOURCES]: { disabled: false },
+      [storageSteps.STORAGE_QUEUES]: { disabled: false },
+      [storageSteps.STORAGE_TABLES]: { disabled: false },
       [STEP_RM_COMPUTE_NETWORK_RELATIONSHIPS]: { disabled: false },
       [STEP_RM_AUTHORIZATION_ROLE_ASSIGNMENTS]: { disabled: false },
       [STEP_RM_AUTHORIZATION_ROLE_ASSIGNMENT_PRINCIPAL_RELATIONSHIPS]: {