+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Added `networkRuleSetDefaultAction` and `networkRuleSetBypass` properties to
+  `azure_storage_account`.
+
 ### Changed
 
 - [#230](https://github.com/JupiterOne/graph-azure/issues/230) - Change job log